Update README.md
Simple changes. Also, 3 years is being too humble and inaccurate. I have close to 2.5 years with OOP languages (Java, C#) professionally. 4.5 years in education, and 2 years personal use. 

Education and personal use may not be the most extensive use of the language, but I have been with Java for a long time.